+++ src/main/java/com/example/social_network_fpt_be/repository/CommentRepository.java
@@ -8,10 +8,10 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CommentRepository extends JpaRepository<Comment, Integer> {
+public interface CommentRepository extends JpaRepository<Comment, Long> {
 
     @Query("SELECT c FROM Comment c WHERE c.id_comment_father = ?1")
-    List<Comment> getCommentByFather(Integer id);
+    List<Comment> getCommentByFather(Long id);
 
     @Query("SELECT c FROM Comment c WHERE c.id_comment_father is null")
     List<Comment> getCommentRoot();
